Change: t1 values so that they are now doubles, add a boolean field which checks whether the m10 bolts are required

+++ ContourPlateBridge/SmartContourPlate.cs
@@ -35,7 +35,7 @@ namespace ContourPlateBridge
         private List<ToleranceReport> _tolerances;
         private int lastTwoDigits;
 
-        public SmartContourPlate(Model model, double xOrigin, double yOrigin, int profile, int t1, int t2, int t3, int t4, double aWidth, double bLength, string bearingMark, List<ToleranceReport> _tolerances, int lastTwoDigits)
+        public SmartContourPlate(Model model, double xOrigin, double yOrigin, int profile, double t1, double t2, double t3, double t4, double aWidth, double bLength, string bearingMark, List<ToleranceReport> _tolerances, int lastTwoDigits)
         {
             this.model = model;
             this.xOrigin = xOrigin;
@@ -90,7 +90,7 @@ namespace ContourPlateBridge
             contourPlate.AssemblyNumber.Prefix = this.name;
             contourPlate.AssemblyNumber.StartNumber = this.lastTwoDigits;
             
-            contourPlate.Insert();
+            contourPlate.Insert();            
 
             insertHorizontalBolts(contourPlate, horizontalBottomBoltOrigin());
             insertHorizontalBolts(contourPlate, horizontalTopBoltOrigin());
